fixed issue with the variable being 64-bit in one place and 32-bit in another
git-svn-id: https://svn.code.sf.net/p/gridlab-d/code/trunk@2327 54b63ff4-7b51-0410-8bda-9fa96797e505

+++ core/class.h
@@ -154,7 +154,7 @@ typedef enum {
 
 typedef struct s_keyword {
 	char name[32];
-	uint32 value;
+	uint64 value;
 	struct s_keyword *next;
 } KEYWORD;
